su tests: using TEST_P() and TEST_SIZE().
darcs-hash:20061121185719-65a35-2fab31d458a013d437ef7ff469971fb0bce2043a.gz

+++ libsofia-sip-ua/su/test_htable.c
@@ -155,7 +155,7 @@ void zap(context_t *c, entry_t *e)
   su_free(c->c_home, e);
 }
 
-static int count(context_t *c, hash_value_t h)
+static unsigned long count(context_t *c, hash_value_t h)
 {
   entry_t *e, **ee;
   unsigned long n;
